refactor(containers): delete the old file based store
