+++ edgehog-device-runtime-containers/src/lib.rs
@@ -37,7 +37,6 @@ pub(crate) mod events;
 mod properties;
 pub mod requests;
 pub mod service;
-pub mod store;
 
 #[cfg(feature = "mock")]
 mod mock;
